{"schema":"decodex/commit/1","summary":"Modularize contract and research design models","authority":"manual"}
